% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_automata.R
\name{ca_cyclic}
\alias{ca_cyclic}
\title{Cyclic Cellular Automata}
\usage{
ca_cyclic(
  ncolumns,
  nrows,
  nstates,
  neighborhood,
  th,
  iterations,
  animate = FALSE,
  seed = NULL
)
}
\arguments{
\item{ncolumns}{number of columns in grid}

\item{nrows}{number of rows in grid}

\item{nstates}{number of states}

\item{neighborhood}{configuration of neighborhood (x, y, weight)}

\item{th}{threshold, should be proportionate to probability of each state}

\item{iterations}{number of iterations}

\item{animate}{animate iterations? defaults to FALSE}

\item{seed}{seed, defaults to NULL}
}
\value{
A a single long grid matrix or frames of long grid for animation
}
\description{
if number of neighbors with state one unit greater than current is larger than the threshold increment current state by one
}
