% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_automata.R
\name{ca_index}
\alias{ca_index}
\title{Index Cellular Automata}
\usage{
ca_index(ncolumns, nrows, nstates, neighborhood, iterations, seed = NULL)
}
\arguments{
\item{nrows}{number of rows in grid}

\item{nstates}{number of states}

\item{neighborhood}{configuration of neighborhood (x, y, weight)}

\item{iterations}{number of iterations}

\item{seed}{seed, defaults to NULL}
}
\value{
A a single long grid matrix or frames of long grid for animation
}
\description{
the state of the cell acts as an index to neighbor from which the cell's next state is taken.
}
