% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_automata.R
\name{ca_total}
\alias{ca_total}
\title{Totalistic Cellular Automata}
\usage{
ca_total(
  ncolumns,
  nrows,
  nstates,
  init_mode,
  neighborhood,
  iterations,
  animate = FALSE,
  seed = NULL
)
}
\arguments{
\item{ncolumns}{number of columns in grid}

\item{nrows}{number of rows in grid}

\item{init_mode}{mode of initial configuration, "R": random, "M": middle of grid}

\item{neighborhood}{configuration of neighborhood (x, y, weight)}

\item{iterations}{number of iterations}

\item{animate}{animate iterations? defaults to FALSE}

\item{seed}{seed, defaults to NULL}
}
\value{
A a single long grid matrix or frames of long grid for animation
}
\description{
Generate a Totalistic CA with animation option
}
