% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary.R
\name{elementary_multi}
\alias{elementary_multi}
\title{Elementary CA Multi-state}
\usage{
elementary_multi(
  rule_num,
  ncolumns,
  nrows,
  nstates,
  first_mode,
  gen_mode,
  seed = NULL
)
}
\arguments{
\item{rule_num}{Wolfram rule number}

\item{ncolumns}{number of columns in grid}

\item{nrows}{number of rows in grid}

\item{nstates}{number of states}

\item{first_mode}{mode of generation of first row. ("M": mid row on, "R": random)}

\item{gen_mode}{mode of state selection. ("S": state of current cell, "N": sample of cell state and neighbors')}

\item{seed}{seed, defaults to NULL}
}
\value{
A matrix of cells
}
\description{
Generate a multi-state Elementary Cellular Automata by mapping the multi-state cells to binary cells that are used as controls for the state selection
}
