\name{normdata}
\alias{normdata}
\docType{data}
\title{ Example of normalized data }
\description{
Normalized data of data set \code{RG}. It is generated by function \code{\link[CALIB:normalizedata]{normalizeData}}
in the CALIB package by using calibration model parameter \code{\link{parameter}}.
}

\usage{data(normdata)}

\format{
This is a numeric matrix. Row corresponds to unique clone on the arrays and
column corresponds to the two different conditions.  
}

\references{
dataset \code{\link[CALIB:rg]{RG}}.

dataset \code{\link[CALIB:parameter]{parameter}}.

Engelen, K., Naudts, B., DeMoor, B., Marchal, K. (2006) 
A calibration method for estimating absolute expression levels from 
microarray data. Bioinformatics  22: 1251-1258.

Hilson,P.,et al. (2004) Versatile gene-specific sequence tags for 
Arabidopsis functional genomics: transcript profiling and reverse
genetics applications. Genome Res. 14, 2176-2189.
}

\examples{
data(normdata)

plotNormalizedData(normdata,condition = c(1,2))
}

\keyword{datasets}
