\name{RG}
\alias{RG}
\docType{data}

\title{ Experiment Data: RGList\_CALIB Example}

\description{
This \code{\link[CALIB:RGList_CALIB-class]{RGList_CALIB}} object
represents microarray data of two arrays. 
It is generated by function \code{\link[CALIB{read.rg}}
from raw data files.
}

\usage{data(RG)}

\format{
\code{RG} is an \code{RGList_CALIB} object containing the following list components:
\$R,\$G,\$Rb,\$Gb,\cr\$RArea,\$GArea,\$targets,\$source and \$genes. It represents
two microarrays and 19749 clones.
}

\source{
This comes from a publicly avaible dataset, consisting 14 hybridizations.
From this \code{RG}, two out of these fourteen are chosen. The experiment design 
of these two are color flip. Except for cDNA probes, external control spikes 
are also spotted on the arrays. There are 192 ratio controls, 480 calibration 
controls,24 negative controls and 72 utility controls.

For more information, see the reference.
}

\references{
Engelen, K., Naudts, B., DeMoor, B., Marchal, K. (2006) 
A calibration method for estimating absolute expression levels from 
microarray data. Bioinformatics  22: 1251-1258.

Hilson,P.,et al. (2004) Versatile gene-specific sequence tags for 
Arabidopsis functional genomics: transcript profiling and reverse
genetics applications. Genome Res. 14, 2176-2189.
}

\examples{
data(RG)
}

\keyword{datasets}
