\name{spike}
\alias{spike}
\docType{data}

\title{ Experiment Data: SpikeList Example}

\description{
This \code{\link[CALIB:spikelist]{SpikeList}} object contains all the spikes in dataset
\code{\link[CALIB:rg]{RG}}.It is generated by the function\cr \code{\link[CALIB:read.spike]{read.spike}}
from \code{RG} and user-specified concentraion file in the CALIB package.
}

\usage{data(spike)}

\format{
\code{spike} is an \code{SpikeList} object containing the following list components:
\$R,\$G,\$Rb,\$Gb,\cr\$RArea,\$GArea,\$RConc,\$GConc and \$genes. It represents two
microarrays and 600 spikes, including 480 calibration controls, 96 ratio
controls and 24 negative controls.
}

\source{
 For the source information, see the introduction of dataset \code{RG}.
}

\references{
dataset \code{RG}.

Engelen, K., Naudts, B., DeMoor, B., Marchal, K. (2006) 
A calibration method for estimating absolute expression levels from 
microarray data. Bioinformatics  22: 1251-1258.

Hilson,P.,et al. (2004) Versatile gene-specific sequence tags for 
Arabidopsis functional genomics: transcript profiling and reverse
genetics applications. Genome Res. 14, 2176-2189.

}
\examples{
data(spike)

plotSpikeRG(spike,array=1)
}

\keyword{datasets}
