\name{SpikeList-class}
\docType{class}
\alias{SpikeList-class}
\alias{show,SpikeList-method}

\title{ Class "SpikeList" - Spike Intensity and Concentration List }

\description{ A simple list-based class for storing red and green channel foreground and
      background intensities, spot area and concentrations for external control spike on
      spotted microarray. }

\section{Objects from the Class}{
   Objects can be created by calls of the form \code{new("SpikeList", spike)} where spike
   is a list. In the CALIB package, SpikeList objects are normally generated by function 
   \code{\link[CALIB:read.spike]{read.spike}}.
}

\section{List Components}{
    Objects should contain the following list components:
	 \describe{ 
    \item{\code{R}:}{ numeric matrix containing the red(Cy5) foreground intensities
          of all external control spikes on arrays. Rows correspond to spikes and 
          columns to arrays.}

    \item{\code{G}:}{ numeric matrix containing the green(Cy3) foreground intensities
          of all external control spikes on arrays. }

    \item{\code{Rb}:}{ numeric matrix containing the red(Cy5) background intensities
          of all external control spikes on arrays. }

    \item{\code{Gb}:}{ numeric matrix containing the green(Cy3) background intensities
          of all external control spikes on arrays. }

    \item{\code{RArea}:}{ numeric matrix containing the red(Cy5) spot areas of all 
          external control spikes on arrays. }

    \item{\code{GArea}:}{ numeric matrix containing the green(Cy3) spot areas of all 
          external control spikes on arrays. }

    \item{\code{RConc}:}{ numeric matrix containing the red(Cy5) known concentrations
          of all external control spikes on arrays.}

    \item{\code{GConc}:}{ numeric matrix containing the green(Cy3) know concentrations
          of all external control spikes on arrays.}

    \item{\code{genes}:}{ data.frame containing information on spikes spotted on the arrays.
          Should include a character column Name containing names for all the spikes.}
  }
    All of the matrices should have the same dimensions. 
}

\section{Extends}{
Class \code{"list"}, from data part.
Class \code{"vector"}, by class \code{"list"}.
}

\section{Methods}{
This class inherits directly from class List, so any operation appropriate
for lists will work on objects of this class. In addition, SpikeList objects
can be \code{\link[CALIB:subsetting]{subsetted}}, \code{\link[CALIB:cbind]{combined}}
and \code{\link[CALIB:merge]{merged}}. SpikeList objects will return dimensions and
hence functions such as \code{\link[CALIB:dim]{dim}}, \code{\link[CALIB:dimnames]{dimnames}},
\code{\link[base:nrow]{nrow}} and \code{\link[base:ncol]{ncol}} are defined.
Generic method \code{\link[methods]{show}} is applied on \code{SpikeList},
so SpikeList will print in a compact way.

SpikeList objects are used on functions such as \code{\link{estimateParameter}} or on some
other data visualization functions like \code{\link{plotSpikeHI}} in the CALIB package.         
}

\references{ the limma package}

\author{ Hui Zhao}

\seealso{
\code{\link[CALIB:rglist_CAILB]{RGList_CALIB}}.

\code{\link[limma:rglist]{RGList}} in limma package. 
}

\keyword{classes}
