% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary.R
\name{wolfram}
\alias{wolfram}
\title{Wolfram CA binary}
\usage{
wolfram(rule_num, ncolumns, nrows, first_mode, seed = NULL)
}
\arguments{
\item{rule_num}{Wolfram rule number}

\item{ncolumns}{number of columns in grid}

\item{nrows}{number of rows in grid}

\item{first_mode}{mode of generation of first row. ("M": mid row on, "R": random)}

\item{seed}{seed, defaults to NULL}
}
\value{
A matrix of cells
}
\description{
Generate the basic binary Wolfram Cellular Automata
}
