% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary.R
\name{wolfram_total}
\alias{wolfram_total}
\title{Wolfram CA totalistic}
\usage{
wolfram_total(
  rule_num = NULL,
  ncolumns,
  nrows,
  nstates,
  first_mode,
  seed = NULL
)
}
\arguments{
\item{rule_num}{if NULL - random generation of rule, otherwise generated based on # of states}

\item{ncolumns}{number of columns in grid}

\item{nrows}{number of rows in grid}

\item{nstates}{number of states}

\item{first_mode}{mode of generation of first row. ("M": mid row on, "R": random)}

\item{seed}{seed, defaults to NULL}
}
\value{
A matrix of cells
}
\description{
Generate the totalistic Wolfram Cellular Automata. It uses summation rule and it supports multiple states
}
